iproute2: Interim patch for iproute2 json_print.c
This is a temporary patch to send iproute2 lib/json_print.c outputs to a memory buffer,
it will allow iproute2-sysrepo module to access the memory buffer data and later send it to sysrepo operational data store.

This temporary solution will be revoked if one of the following long-term solutions get applied:
a- upstream changes to iproute2 that allows to set the json_print.c output file.
b- sending the complete iproute2 stdout to a memory buffer.

The current interim change involves the following:
[+] iproute2-sysrepo: introduce json_buffer to hold iproute2 show commands json string outputs.
[+] iproute2: patch lib/json_print.c to send outputs to json_buffer instead of stdout.
[+] Makefile: temporary patch to apply iproute2 json_print patch before compiling.

Signed-off-by: Amjad Daraiseh <adaraiseh@okdanetworks.com>