iproute2/json: dump into mem buf instead of stdout
It is a link time patching instead of patching the C code of
iproute2/lib/json_print.c

A C patch was tried by this pull request too:
  https://github.com/okda-networks/iproute2-sysrepo/pull/34

It is a dirty option. The proper solution would be an upstream
patch of iproute2 in order to support either a stdout output or
a mem stream output.